+++ HydroNotifier.FunctionApp/Core/WebScraper.cs
@@ -34,7 +34,7 @@ namespace HydroNotifier.FunctionApp.Core
             var lastValueRow = rows[1];
             var values = lastValueRow.SelectNodes("td");
             var date = values[0].InnerText;
-            var flowLitersPerSecond = decimal.Parse(values[2].InnerText, NumberStyles.Any, CultureInfo.InvariantCulture) * 1000.0M;
+            var flowLitersPerSecond = double.Parse(values[2].InnerText, NumberStyles.Any, CultureInfo.InvariantCulture) * 1000.0;
             _log.LogInformation($"Name = '{_query.Name}', Date = '{date}', FlowInLitersPerSecond = '{flowLitersPerSecond}'");
 
             return new HydroData(_query.Name, date, flowLitersPerSecond);
