+++ HydroNotifier.FunctionApp/Utils/ApplicationInsightsTelemetry.cs
@@ -8,9 +8,9 @@ namespace HydroNotifier.FunctionApp.Utils
     {
         private readonly TelemetryClient _tc = new TelemetryClient(TelemetryConfiguration.CreateDefault());
 
-        public void TrackMetric(string name, decimal value)
+        public void TrackMetric(string name, double value)
         {
-            _tc.TrackMetric(name, (double)value);
+            _tc.TrackMetric(name, value);
         }
 
         public void TrackEvent(string eventName, IDictionary<string, string> properties = null, IDictionary<string, double> metrics = null)
