+++ HydroNotifier.FunctionApp/Utils/ITelemetry.cs
@@ -4,7 +4,7 @@ namespace HydroNotifier.FunctionApp.Utils
 {
     public interface ITelemetry
     {
-        void TrackMetric(string name, decimal value);
+        void TrackMetric(string name, double value);
         void TrackEvent(string eventName, IDictionary<string, string> properties = null, IDictionary<string, double> metrics = null);
     }
 }
